- changed many data types from "unsigned int" to "size_t" ->must also be adapted in the ROS-node - added additional functionality to stereo_pose_refinement - THIS CODE MIGHT NOT COMPILE! Use the version before the last version. git-svn-id: https://redmine.ait.ac.at/sas-svn/navigation/trunk/matching@417 5a9f4643-6f94-47ed-9025-08a63a7baffb

+++ source/poselib/include/poselib/pose_linear_refinement.h
@@ -39,11 +39,11 @@ namespace poselib
 		cv::InputOutputArray E,
 		cv::InputOutputArray mask,
 		int refineMethod,//a combination of poselib::RefinePostAlg
-		unsigned int & nr_inliers,
+		size_t & nr_inliers,
 		cv::InputOutputArray R = cv::noArray(),
 		cv::OutputArray t = cv::noArray(),
 		double th = 0.8,
-		unsigned int num_iterative_steps = 4,
+		size_t num_iterative_steps = 4,
 		double threshold_multiplier = 2.0,
 		double pseudoHuberThreshold_multiplier = 0.1,
 		double maxRelativeInlierCntLoss = 0.15);
